+++ src/gdalraster.h
@@ -167,6 +167,8 @@ class GDALRaster {
 	bool setColorTable(int band, Rcpp::RObject &col_tbl,
 			std::string palette_interp);
 	
+	SEXP getDefaultRAT(int band) const;
+	
 	void flushCache();
 	
 	int getChecksum(int band, int xoff, int yoff, int xsize, int ysize) const;